добавил поддержку темы в StoryBook
